+++ common/src/main/java/net/mehvahdjukaar/amendments/common/recipe/RecipeUtils.java
@@ -62,7 +62,8 @@ public class RecipeUtils {
 
 
     public static ItemStack simulateCrafting(Level level, ItemStack dye, ItemStack playerItem, boolean surround) {
-        DummyContainer container = new DummyContainer(dye.copy(), playerItem.copy(), surround);
+        DummyContainer container =  surround ? DummyContainer.surround(dye.copy(), playerItem.copy()) :
+                DummyContainer.of(dye.copy(), playerItem.copy());
         var recipes = level.getRecipeManager().getRecipesFor(RecipeType.CRAFTING, container, level);
         for (var r : recipes) {
             ItemStack recolored = r.assemble(container, level.registryAccess());
@@ -77,92 +78,4 @@ public class RecipeUtils {
         return null;
     }
 
-    private static class DummyContainer implements CraftingContainer {
-
-        private final List<ItemStack> items = new ArrayList<>();
-        private final boolean surround;
-
-        public DummyContainer(ItemStack dye, ItemStack toRecolor, boolean surround) {
-            this.surround = surround;
-            if (surround) {
-                items.add(toRecolor);
-                items.add(toRecolor);
-                items.add(toRecolor);
-
-                items.add(toRecolor);
-                items.add(dye);
-                items.add(toRecolor);
-
-                items.add(toRecolor);
-                items.add(toRecolor);
-                items.add(toRecolor);
-            } else {
-                items.add(dye);
-                items.add(toRecolor);
-            }
-        }
-
-        @Override
-        public int getContainerSize() {
-            return items.size();
-        }
-
-        @Override
-        public boolean isEmpty() {
-            return items.isEmpty();
-        }
-
-        @Override
-        public ItemStack getItem(int slot) {
-            if (slot >= this.getContainerSize()) return ItemStack.EMPTY;
-            return items.get(slot);
-        }
-
-        @Override
-        public ItemStack removeItem(int slot, int amount) {
-            return null;
-        }
-
-        @Override
-        public ItemStack removeItemNoUpdate(int slot) {
-            return null;
-        }
-
-        @Override
-        public void setItem(int slot, ItemStack stack) {
-        }
-
-        @Override
-        public void setChanged() {
-        }
-
-        @Override
-        public boolean stillValid(Player player) {
-            return true;
-        }
-
-        @Override
-        public void clearContent() {
-        }
-
-        @Override
-        public int getWidth() {
-            return surround ? 3 : 2;
-        }
-
-        @Override
-        public int getHeight() {
-            return surround ? 3 : 2;
-        }
-
-        @Override
-        public List<ItemStack> getItems() {
-            return items;
-        }
-
-        @Override
-        public void fillStackedContents(StackedContents helper) {
-
-        }
-    }
 }