+++ common/src/main/java/net/mehvahdjukaar/amendments/events/behaviors/SkullCandleConversion.java
@@ -75,7 +75,7 @@ public class SkullCandleConversion implements ItemUseOnBlock {
 
                 InteractionResult result = InteractEvents.replaceSimilarBlock(b,
                         player, stack, pos, world, state, SoundType.CANDLE,
-                        true, SkullBlock.ROTATION, WallSkullBlock.FACING);
+                        true,true, SkullBlock.ROTATION, WallSkullBlock.FACING);
 
                 if (result.consumesAction()) {
                     if (world.getBlockEntity(pos) instanceof CandleSkullBlockTile tile) {