+++ common/src/main/java/net/mehvahdjukaar/amendments/mixins/AbstractCauldronBlockMixin.java
@@ -2,8 +2,12 @@ package net.mehvahdjukaar.amendments.mixins;
 
 import com.llamalad7.mixinextras.injector.ModifyReturnValue;
 import com.llamalad7.mixinextras.sugar.Local;
+import net.mehvahdjukaar.amendments.common.block.BoilingWaterCauldronBlock;
+import net.mehvahdjukaar.amendments.common.block.LiquidCauldronBlock;
 import net.mehvahdjukaar.amendments.configs.CommonConfigs;
 import net.mehvahdjukaar.amendments.events.behaviors.CauldronConversion;
+import net.mehvahdjukaar.moonlight.api.fluids.BuiltInSoftFluids;
+import net.mehvahdjukaar.moonlight.api.fluids.SoftFluidStack;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
@@ -33,6 +37,15 @@ public class AbstractCauldronBlockMixin extends Block {
         if (original == InteractionResult.PASS && this == Blocks.CAULDRON && CommonConfigs.LIQUID_CAULDRON.get()) {
             return CauldronConversion.convert(state, pos, level, player, hand, stack);
         }
+        BlockState newState = level.getBlockState(pos);
+        if (newState.getBlock() == Blocks.CAULDRON) {
+            boolean isFire = LiquidCauldronBlock.shouldBoil(level.getBlockState(pos.below()),
+                    new SoftFluidStack(BuiltInSoftFluids.WATER.getHolder()));
+            if (isFire) {
+                level.setBlockAndUpdate(pos, newState.setValue(BoilingWaterCauldronBlock.BOILING, true));
+            }
+        }
+
         return original;
     }